requirements.txt
pyyaml>=6.0
jsbeautifier>=1.14.0